Remove image from the published crates.io package
During a dependency review I noticed that imgref includes an image in
the published package. This file is not required for building imgref
as dependency and make it a bit harder to review the code.

This commit introduces a `include` directive in the `Cargo.toml` file to
explicitly include only this file required to build `imgref`.
Additionally it changes the Readme to use an absulute link to the image
instead so that it continues to be rendered correctly everywhere. This
allows to exclude the image and also reduces the size of the
published crate from 15 files, 119.0KiB (64.7KiB compressed) to
11 files, 73.6KiB (18.4KiB compressed) which results in a 65GB/Month
traffic reduction for crates.io assuming the current 1.4 million
downloads and the difference in the compressed package size.